feat: structuring endpoints
