--- services/agent_history/app/api/routers/enpoints.py
@@ -1,157 +1,0 @@
-from fastapi import APIRouter, HTTPException, Response, status
-
-from app.api.schemas import AgentResponse, Discussion, SystemMessage, Tool
-from app.api.deps import agent_response_manager
-from app.logs import get_logger
-
-logger = get_logger(__name__)
-
-router = APIRouter(tags=["agents"])
-
-@router.get(
-    "/discussions/{discussion_id}",
-    summary="Получить все ответы агентов в дискуссии",
-    response_model=Discussion,
-    status_code=200,
-    responses={
-        200: {"description": "Ответы агентов"},
-        404: {"description": "Ответы агентов или дискуссия не найдены"}
-    }
-)
-async def get_response_in_discussion(
-    discussion_id: str
-):
-    """Получить ответы агентов"""
-    try:
-
-        events = agent_response_manager.get_discussion_history(
-            discussion_id=discussion_id
-        )
-
-        if events is None:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Дискуссия '{discussion_id}' не найдена"
-            )
-
-        return Discussion(
-            discussion_id=discussion_id,
-            events=events
-        )
-    except Exception as e:
-        logger.error(f"Ошибка при получении всей информации о дискуссии: {e}")
-        raise HTTPException(status_code=500, detail=f"Ошибка при получении всей информации о дискуссии: {str(e)}")
-
-
-@router.delete(
-    "/discussions/{discussion_id}",
-    summary="Удалить дискуссию",
-    status_code=204,
-    responses={
-        204: {"description": "Дискуссия успешно удалена"},
-        404: {"description": "Дискуссия не найдена"}
-    }
-)
-async def delete_discussion(
-    discussion_id: str
-):
-    """Удалить дискуссию"""
-    try:
-
-        success = agent_response_manager.delete_discussion(
-            discussion_id=discussion_id,
-        )
-
-        if not success:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Дискуссия '{discussion_id}' не найдена"
-            )
-
-        return Response(
-            status_code=status.HTTP_204_NO_CONTENT
-        )
-    except Exception as e:
-        logger.error(f"Ошибка при удалении дискуссии: {e}")
-        raise HTTPException(status_code=500, detail=f"Ошибка при удалении дискуссии: {str(e)}")
-
-@router.post(
-    "/discussions/{discussion_id}/responses",
-    summary="Сохранить ответ агента",
-    status_code=201,
-    responses={
-        201: {"description": "Ответ агента успешно сохранён"},
-    }
-)
-async def save_response(
-    discussion_id: str,
-    response: AgentResponse
-):
-    """Сохранить новый ответ агента"""
-    try:
-
-        agent_response_manager.save_response(
-            discussion_id=discussion_id,
-            response=response
-        )
-
-        return Response(
-            status_code=status.HTTP_201_CREATED
-        )
-    except Exception as e:
-        logger.error(f"Ошибка при сохранении ответа агента: {e}")
-        raise HTTPException(status_code=500, detail=f"Ошибка при сохранении ответа агента: {str(e)}")
-
-@router.post(
-    "/discussions/{discussion_id}/system_messages",
-    summary="Сохранение сообщения от системы",
-    status_code=201,
-    responses={
-        201: {"description": "Сообщение системы успешно сохранено"},
-    }
-)
-async def post_system_message(
-    discussion_id: str,
-    message: SystemMessage
-):
-    """Сохранить сообщение системы"""
-    try:
-
-        agent_response_manager.save_system_mes(
-            discussion_id=discussion_id,
-            message=message
-        )
-
-        return Response(
-            status_code=status.HTTP_201_CREATED
-        )
-    except Exception as e:
-        logger.error(f"Ошибка при сохранении сообщения от системы: {e}")
-        raise HTTPException(status_code=500, detail=f"Ошибка при сохранении сообщения от системы: {str(e)}")
-
-@router.post(
-    "/discussions/{discussion_id}/tool",
-    summary="Сохранение информации об инструменте",
-    status_code=201,
-    responses={
-        201: {"description": "Информацию об инструмене успешно сохранено"},
-    }
-)
-async def post_tool(
-    discussion_id: str,
-    tool_info: Tool
-):
-    """Сохранить информацию об инструменте"""
-    try:
-
-        agent_response_manager.save_tool_info(
-            discussion_id=discussion_id,
-            tool_info=tool_info
-        )
-
-        return Response(
-            status_code=status.HTTP_201_CREATED
-        )
-    except Exception as e:
-        logger.error(f"Ошибка при сохранении информации об инструементе: {e}")
-        raise HTTPException(status_code=500, detail=f"Ошибка при сохранении информации об инструменте: {str(e)}")
