+++ services/agent_history/app/api/routers/info.py
@@ -1,7 +1,10 @@
 from fastapi import APIRouter, HTTPException, status
 from fastapi.responses import JSONResponse
 
-from app.api.schemas import SystemMessageType, HealthResponse, ToolStatus
+from app.api.schemas import (
+    SystemMessageType, HealthResponse, 
+    ToolStatus, AgentStatus
+)
 
 routers = APIRouter(
     prefix="/info", 
@@ -67,6 +70,29 @@ async def get_type_tool():
     """Получаем статусы инструментов"""
     try:
         types = [t.value for t in ToolStatus]
+        return JSONResponse(
+            status_code=status.HTTP_200_OK,
+            content={"types": types}
+        )
+    except Exception as e:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail=f"Неопределённая ошибка: {e}"
+        )
+
+@routers.get(
+    "/status_agent",
+    summary="Статусы агентов",
+    description="Получить список возможных статусов агентов",
+    responses={
+        200: {"description": "Получен список статусов"},
+        500: {"description": "Внутренняя ошибка сервера"},
+    }
+)
+async def get_type_agent():
+    """Получаем статусы агентов"""
+    try:
+        types = [t.value for t in AgentStatus]
         return JSONResponse(
             status_code=status.HTTP_200_OK,
             content={"types": types}
