+++ services/agent_history/app/api/schemas/discussion.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel, Field
 
 class Discussion(BaseModel):
     discussion_id: str = Field(..., description="ID дискуссии")
-    events: List[dict] = Field(..., description="Событие. Сообщение от агента или от системы")
+    events: List[dict] = Field(..., description="Событие. Сообщение от агента/системы/инструмента")
